+++ main.py
@@ -1,12 +1,14 @@
 from downloader.fiction import Fiction
-from downloader.chapter_fetcher import ChapterFetcher
+from downloader.fetcher import ChapterFetcher
+from downloader.renderer import HTMLRenderer
 
 fiction = Fiction(
-    "Mother of Learning",
-    21220,
-    "mother-of-learning",
-    "301778/1-good-morning-brothe",
+    "Delve",
+    25225,
+    "delve",
+    "368012/001-woodland",
 )
 fetcher = ChapterFetcher(fiction)
 chapters = fetcher.fetch(up_to_chapter=2)
-print("done")
+renderer = HTMLRenderer()
+renderer.save_to_file(fiction, chapters)
